+++ ChapaInAppPurchase/src/main/java/com/yosef/chapainapppurchase/payment_type/BasicPayment.java
@@ -8,7 +8,7 @@ import com.yosef.chapainapppurchase.PaymentType;
  */
 public class BasicPayment extends PaymentType {
 
-    public BasicPayment(float amount) {
+    public BasicPayment(double amount) {
         super(amount);
     }
